+++ rust/operator-binary/src/main.rs
@@ -47,6 +47,7 @@ mod hdfs_controller;
 mod operations;
 mod product_logging;
 mod security;
+mod service;
 
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
